* Headers/Additions/GNUstepGUI/GSToolbarView.h 	* Source/GSToolbarView.m 	Change draggedItemIndex to NSUInteger.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@40270 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSToolbarView.h
@@ -60,8 +60,8 @@ typedef enum {
   float _heightFromLayout;
 }
 
-+ (int) draggedItemIndex;
-+ (void) setDraggedItemIndex:(int)sourceIndex;
++ (NSUInteger) draggedItemIndex;
++ (void) setDraggedItemIndex:(NSUInteger)sourceIndex;
 
 - (id) initWithFrame: (NSRect)frame;
